add envs folder
