--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "rlkit"]
-	path = rlkit
-	url = https://github.com/natolambert/rlkit
